chore: Movido publish para ExchangeService
