+++ LagRabbitMqManagerToolkit.Tests/Integrations/QueueRequestTests.cs
@@ -8,12 +8,14 @@ namespace LagRabbitMqManagerToolkit.Tests.Integrations
     public class QueueRequestTests
     {
         private readonly IQueueService _queueRequest;
+        private readonly IExchangeService _exchangeRequest;
 
         public QueueRequestTests()
         {
             var settings = RabbitSettings.Default();
 
             _queueRequest = new QueueService(settings);
+            _exchangeRequest = new ExchangeService(settings);
         }
 
         [Fact]
@@ -34,7 +36,7 @@ namespace LagRabbitMqManagerToolkit.Tests.Integrations
 
             var payload = "<teste></teste>";
 
-            await _queueRequest.PublishAsync("/", "teste", properties, payload);
+            await _exchangeRequest.PublishAsync("/", "teste", properties, payload);
         }
     }
 }
