+++ LagRabbitMqManagerToolkit/Services/Interfaces/IQueueService.cs
@@ -9,7 +9,5 @@ namespace LagRabbitMqManagerToolkit.Services.Interfaces
         Task<Queue?> GetAsync(string vHost, string queue);
 
         Task<IList<Message>> GetMessagesAsync(string vHost, string queue, int take = 200);
-
-        Task PublishAsync(string vHost, string queue, Dictionary<string, string> properties, string payload, string exchange = "amq.default", string encoding = "string");
     }
 }
