+++ LagRabbitMqManagerToolkit/Services/QueueService.cs
@@ -52,27 +52,5 @@ namespace LagRabbitMqManagerToolkit.Services
 
             return await RabbitRequestExtensions.Get<List<Queue>>(url, token) ?? [];
         }
-
-        public async Task PublishAsync(string vHost, string queue, Dictionary<string, string> properties, string payload, string exchange = "amq.default", string encoding = "UTF-8")
-        {
-            var endpoint = RabbitEndpoints.PublishMessage(vHost, exchange);
-
-            var url = new Uri(new Uri(_settings.Url), endpoint);
-
-            var token = RabbitRequestExtensions.BasicToken(_settings);
-
-            var body = new
-            {
-                properties,
-
-                routing_key = queue,
-
-                payload,
-
-                payload_encoding = encoding
-            };
-
-            await RabbitRequestExtensions.Post(url, token, body);
-        }
     }
 }
